+++ UniVgo/Runtime/Utils/ShaderName.cs
@@ -1,0 +1,45 @@
+﻿// ----------------------------------------------------------------------
+// @Namespace : UniVgo
+// @Class     : ShaderName
+// ----------------------------------------------------------------------
+namespace UniVgo
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class ShaderName
+    {
+        /// <summary></summary>
+        public const string UniGLTF_StandardVColor = "UniGLTF/StandardVColor";
+
+        /// <summary></summary>
+        public const string Unlit_Color = "Unlit/Color";
+
+        /// <summary></summary>
+        public const string Unlit_Texture = "Unlit/Texture";
+
+        /// <summary></summary>
+        public const string Unlit_Transparent = "Unlit/Transparent";
+
+        /// <summary></summary>
+        public const string Unlit_Transparent_Cutout = "Unlit/Transparent Cutout";
+
+        /// <summary></summary>
+        public const string UniGLTF_UniUnlit = "UniGLTF/UniUnlit";
+
+        /// <summary></summary>
+        public const string VRM_UnlitTexture = "VRM/UnlitTexture";
+
+        /// <summary></summary>
+        public const string VRM_UnlitTransparent = "VRM/UnlitTransparent";
+
+        /// <summary></summary>
+        public const string VRM_UnlitCutout = "VRM/UnlitCutout";
+
+        /// <summary></summary>
+        public const string VRM_UnlitTransparentZWrite = "VRM/UnlitTransparentZWrite";
+
+        /// <summary></summary>
+        public const string VRM_MToon = "VRM/MToon";
+    }
+}