+++ UniVgo/Runtime/VgoVersion.cs
@@ -13,12 +13,12 @@ namespace UniVgo
         public const int MAJOR = 0;
 
         /// <summary>Minor</summary>
-        public const int MINOR = 5;
+        public const int MINOR = 6;
 
         /// <summary>Patch</summary>
         public const int PATCH = 0;
 
         /// <summary>Version</summary>
-        public const string VERSION = "0.5.0";
+        public const string VERSION = "0.6.0";
     }
 }